% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not in operator}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{vector or list}

\item{y}{vector or list}
}
\value{
A logical vector indicating if elements of \code{x}
are NOT contained in \code{y}.
}
\description{
A negation of the \code{\%in\%} operator.
}
\examples{
1:5 \%!in\% c(1, 3, 5)

}
