% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_save.R
\name{make_pptx}
\alias{make_pptx}
\title{Create an empty PowerPoint file}
\usage{
make_pptx(path)
}
\arguments{
\item{path}{File path to write the .pptx file.}
}
\value{
Invisibly returns \code{path}.
}
\description{
Create a new PowerPoint document and write it to disk.
}
\examples{
\dontrun{
make_pptx("figure.pptx")
}
}
