% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancombc2_plot.R
\name{lfc_num}
\alias{lfc_num}
\title{Plot ANCOM-BC2 results}
\usage{
lfc_num(
  data_prim,
  variable,
  sig_color = "#0FA060",
  nonsig_color = "black",
  title_prefix = "Log fold changes as one unit increase of"
)
}
\arguments{
\item{data}{ANCOM-BC2 result list}

\item{var}{variable name}

\item{title}{plot title}

\item{col_sig}{color for significant taxa}

\item{non_sig}{color for non-significant taxa}

\item{method}{analysis method}
}
\value{
ggplot object
}
\description{
Plot ANCOM-BC2 results
}
